+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # Version 3
 
+## Version 3.4.0
+### 25-Aug-2025
+- Added App Auto-Patch Script Self Update functionality
+- Standardize timestamp format and use actual timezones instead of hard-coded UTC. Cleand up and adjusted NextAutoLaunch format to use date datatype (#152)
+- Added check for appName in Installomator label to populate the correct app name to improve app detection
+- Added logic to check for appCustomVersion in Installomator label to pull correct version of installed apps
+- Fixed logic to clear targetDir variable when scrubbing Installomator label fragments
+
 ## Version 3.3.0
 ### 21-Aug-2025
 - Added functions to write patching receipts into the App Auto Patch management folder. Receipts are used to report success/failure on app patching